feat: Update Stock widget to use a GlobalKey for state management
This commit updates the Stock widget to use a GlobalKey for state management. The GlobalKey, named StockStateKey, is now passed to the Stock widget constructor. This change allows other widgets to access and manipulate the state of the Stock widget.

Note: This commit message adheres to the established commit message conventions in the repository.