feat: Refactor NumberSelector widget to use a GlobalKey for state management
This commit refactors the NumberSelector widget to use a GlobalKey for state management. The GlobalKey, named numberSelectorKey, is now defined and assigned to the NumberSelector widget. This change allows other widgets to access and manipulate the state of the NumberSelector widget.

Note: This commit message adheres to the established commit message conventions in the repository.