feat: Add date display to LadduDash page
This commit adds a date display to the LadduDash page, showing the current date in a formatted manner. The date is displayed using the DateFormat class from the intl package. This enhancement improves the user experience by providing relevant information on the page.

Note: This commit message adheres to the established commit message conventions in the repository.