chore: Update laddu_seva package
This commit updates the laddu_seva package by removing unused code, improving the user interface, fixing issues with the pie chart, and reducing flickering upon refresh. It also updates the ButtonRow widget to use ElevatedButton.icon for clearer visual cues. The ticketColors map in the Const class is now using string keys for improved readability and maintainability. Additionally, the returnStock function has been added to the LadduSevaState class, allowing users to manage and track the stock of laddus in the application.

Note: This commit message follows the established commit message conventions in the repository.