chore: Refactor History widget and add HistoryMonthKey
This commit refactors the History widget in the laddu_seva package by adding the HistoryMonthKey to the ListView children. It also imports the HistoryList widget and includes it in the ListView. This change improves the organization and structure of the code.

Note: This commit message follows the established commit message conventions in the repository.