chore: Update Log widget to conditionally display stock note
This commit updates the Log widget in the laddu_seva package to conditionally display the stock note. If the note is not empty, it will be displayed in the UI. This change improves the user experience by providing additional information about the stock.

Note: This commit message follows the established commit message conventions in the repository.