refactor: Update endSession calculation in _HistoryListState
This commit updates the calculation of the endSession variable in the _HistoryListState class. Previously, it was set to the last timestamp in the dists list, but now it is set to the session variable. This change ensures that the endSession value accurately represents the session time.

Note: This commit message follows the established commit message conventions in the repository.